docs: add repository map reference to README.md
Adding a link to the high-level diagrams including module, library dependency and others (https://sourcespy.com/github/vodafonehellov2xworldandroid/). Built directly from source and updated on schedule. Intended to simplify developer's introduction to the project.

In the spirit of transparency - I am the author of the diagrams. Hope contributors find it useful.